+++ pytest_httpx/_pretty_print.py
@@ -1,6 +1,6 @@
 from typing import Union
 
-import httpx
+from pytest_httpx._compat import httpx
 
 from pytest_httpx._httpx_internals import _proxy_url
 from pytest_httpx._request_matcher import _RequestMatcher