+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - `match_params` is now handling `bool`, `int` and `float` values in addition to `str`. Only str values were previously expected.
 - A meaningful error will now be returned when a callback that does not return an `httpx.Response` is called.
+### Added
+- Support for httpx2 is implemented via a httpx/httpx2 compatibility module.
 
 ## [0.36.2] - 2026-04-09
 ### Fixed